Make dealing with uint32_t a little nicer; add u prefix in places

+++ src_pksv/lang_decompiler.h
@@ -9,7 +9,7 @@
 #include "lang_parsers_types.h"
 #include "lang_load.h"
 
-void decompile_all(FILE *input_file, size_t start_offset,
+void decompile_all(FILE *input_file, uint32_t start_offset,
                    const struct language_def *start_language,
                    struct language_cache *language_cache,
                    struct parser_cache *parser_cache, FILE *output_file);
